Continue backend API simplification
